smart-invoice-ai/requirements.txt
streamlit==1.28.1
pandas==2.1.3
numpy==1.24.4
pillow==10.1.0
sqlalchemy==2.0.23
python-multipart==0.0.6
pytesseract==0.3.10
pdf2image==1.17.0
opencv-python-headless==4.8.1.78
plotly==5.17.0